Refactor resources controller and tests

+++ app/controllers/resources_controller.rb
@@ -1,20 +1,24 @@
 class ResourcesController < ApplicationController
   def index
-    render json: Resource.all
+    render_success(:ok, Resource.order(name: :asc))
   end
 
   def create
     resource = Resource.create(resource_params)
-    render json: resource
+    return render_success(:created, resource) if resource.valid?
+    render_error(:unprocessable_entity, resource.errors)
   end
 
   def update
     current_resource.update_attributes(resource_params)
-    render json: current_resource
+    return render_success(:ok, current_resource) if current_resource.valid?
+    render_error(:unprocessable_entity, current_resource.errors)
   end
 
   def destroy
+    return render_error(:not_found) if current_resource.nil?
     current_resource.destroy
+    render_success(:no_content)
   end
 
   private
@@ -24,6 +28,6 @@ class ResourcesController < ApplicationController
   end
 
   def current_resource
-    @current_resource ||= Resource.find(params[:id])
+    @current_resource ||= Resource.find_by_id(params[:id])
   end
 end
